+++ MaldaLang.Tests/InterpretTranspilePairTests.cs
@@ -648,6 +648,22 @@ public class InterpretTranspilePairTests
     }
 
     [Fact]
+    public void ImplicitField_BeatsOuterVar_SameStdout()
+    {
+        InterpretTranspilePair.AssertSameFromSource(
+            """
+            var x = 10;
+            class Point(x, y) {
+                function total() {
+                    return x + y;
+                }
+            }
+            io.print(new Point(2, 3).total());
+            """,
+            "implicit-field-beats-outer-var");
+    }
+
+    [Fact]
     public void PrimaryConstructor_BracelessMethod_SameStdout()
     {
         InterpretTranspilePair.AssertSameFromSource(